Implemented a Renderer Application that can facilitate the rendering of Renderable Components efficiently. Additionally, Lighting (Phong Shading) has been implemented with 8 lights being available to use.

+++ octvis/public/Camera.h
@@ -134,7 +134,7 @@ namespace octvis {
             };
 
             // Converting Degrees to Radians
-            constexpr auto to_radians = [](float degrees) -> float {
+            constexpr auto to_radians = [](double degrees) -> double {
                 return degrees * (PI / 180.0F);
             };
